Add ThreadSafeObservableProperty
